[ refactor ] Splitting InteractionTop.hs
